src: Changing all remaining i32 variables to usize type.
Removing "as usize" casts that are no longer needed using:

#![warn(trivial_numeric_casts)]

+++ src/control.rs
@@ -26,5 +26,5 @@ pub fn control(control: &Control) {
         println!("    aggressive absorption:  no");
     }
 
-    print!("    size of AMD integer: {}\n\n", size_of::<i32>());
+    print!("    size of AMD integer: {}\n\n", size_of::<isize>());
 }
